@@ -3,75 +3,59 @@
   "description": "vertical timeline bar with bookmarks support",
   "author": "Jesus Trujillo",
   "user": "yeyus",
-  "version": "0.1.2",
+  "version": "0.2.0",
   "scripts": {
-    "start": "webpack-dev-server",
-    "test": "karma start",
-    "test:tdd": "karma start --auto-watch --no-single-run",
-    "test:lint": "eslint . --ext .js --ext .jsx",
-    "gh-pages": "webpack",
-    "gh-pages:deploy": "gh-pages -d gh-pages",
-    "gh-pages:stats": "webpack --profile --json > stats.json",
-    "dist": "webpack",
-    "dist:min": "webpack",
-    "dist:modules": "babel ./src --out-dir ./dist-modules",
-    "preversion": "npm run test && npm run test:lint && npm run dist && npm run dist:min && git commit --allow-empty -am \"Update dist\"",
-    "prepublish": "npm run dist:modules",
+    "start": "catalog start docs",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "test:watch": "jest --watch",
+    "test:lint": "eslint . --ignore-path .gitignore",
+    "gh-pages": "catalog build docs",
+    "gh-pages:deploy": "gh-pages -d docs/build",
+    "dist:es6": "del-cli ./dist-es6 && cross-env BABEL_ENV=es6 babel ./src --out-dir ./dist-es6",
+    "dist:modules": "del-cli ./dist-modules && cross-env BABEL_ENV=modules babel ./src --out-dir ./dist-modules",
+    "preversion": "npm run test",
+    "prepublishOnly": "npm run dist:es6 && npm run dist:modules",
     "postpublish": "npm run gh-pages && npm run gh-pages:deploy",
     "postinstall": "node lib/post_install.js"
   },
   "main": "dist-modules",
+  "module": "dist-es6",
+  "jsnext:main": "dist-es6",
   "devDependencies": {
-    "babel-cli": "^6.4.0",
-    "babel-core": "^6.4.0",
-    "babel-eslint": "^5.0.0-beta6",
-    "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.1",
-    "babel-register": "^6.4.3",
-    "chai": "^3.4.1",
-    "clean-webpack-plugin": "^0.1.6",
-    "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-loader": "^1.2.0",
-    "eslint-plugin-react": "^3.11.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
-    "gh-pages": "^0.8.0",
-    "git-prepush-hook": "^1.0.1",
-    "highlight.js": "^9.1.0",
-    "html-webpack-plugin": "^1.7.0",
-    "isparta-loader": "^1.0.0",
-    "json-loader": "^0.5.4",
-    "karma": "^0.13.19",
-    "karma-chai": "^0.1.0",
-    "karma-chai-sinon": "^0.1.5",
-    "karma-coverage": "^0.5.3",
-    "karma-mocha": "^0.2.1",
-    "karma-phantomjs-launcher": "^0.2.3",
-    "karma-sinon": "^1.0.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.23",
-    "karma-webpack": "^1.7.0",
-    "markdown-to-react-components": "^0.2.1",
-    "mocha": "^2.3.4",
-    "phantomjs": "^1.9.19",
-    "phantomjs-polyfill": "0.0.1",
-    "purecss": "^0.6.0",
-    "react": "^0.14.6",
-    "react-addons-test-utils": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-ghfork": "^0.3.2",
-    "sinon": "^1.17.2",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.2",
+    "babel-jest": "^22.4.1",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "catalog": "^3.4.0",
+    "chai": "^4.1.2",
+    "cross-env": "^5.1.3",
+    "del-cli": "^1.1.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "eslint": "^4.18.2",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0",
+    "gh-pages": "^1.1.0",
+    "git-prepush-hook": "^1.0.2",
+    "jest": "^22.4.2",
+    "purecss": "^1.0.0",
+    "react": "^16.2.0",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dom": "^16.2.0",
+    "react-github-corner": "^2.1.0",
+    "rimraf": "^2.6.2",
     "sync-exec": "^0.6.2",
-    "system-bell-webpack-plugin": "^1.0.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.12.11",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-merge": "^0.7.3"
+    "webpack": "^3.10.0"
+  },
+  "peerDependencies": {
+    "react": ">= 0.11.2 < 16.0.2"
   },
   "repository": {
     "type": "git",
